+++ app/services/scoring.js
@@ -102,6 +102,21 @@ export const getMaxScore = (item) => {
   }, valueConstraints.multipleChoice ? 0 : -oo);
 }
 
+const isValueInRange = (value, lookupInfo) => {
+  if (!lookupInfo || lookupInfo == value) {
+    return true;
+  }
+
+  const matched = lookupInfo.match(/^([\d.]+)\s*[-~]\s*([\d.]+)$/);
+
+  if (matched) {
+    value = parseInt(value);
+
+    return !isNaN(value) && value >= Number(matched[1]) && value <= Number(matched[2]);
+  }
+  return false;
+};
+
 export const getScoreFromLookupTable = (responses, jsExpression, items, lookupTable) => {
   let scores = [];
 
@@ -112,37 +127,59 @@ export const getScoreFromLookupTable = (responses, jsExpression, items, lookupTa
   }
 
   let subScaleScore = evaluateScore(jsExpression, items, scores);
-  if (!lookupTable) {
-    return subScaleScore;
+
+  if (lookupTable) {
+    const age = responses[items.findIndex(item => item.variableName === 'age_screen')];
+    const gender = responses[items.findIndex(item => item.variableName === 'gender_screen')].value ? 'F' : 'M';
+
+    for (let row of lookupTable) {
+      if ( 
+        isValueInRange(subScaleScore, row.rawScore) && 
+        isValueInRange(age, row.age) &&
+        isValueInRange(gender, row.sex.toUpperCase())
+      ) {
+        return {
+          tScore: Number(row.tScore),
+          outputText: row.outputText
+        };
+      }
+    }
   }
 
-  const age = responses[items.findIndex(item => item.variableName === 'age_screen')];
-  const gender = responses[items.findIndex(item => item.variableName === 'gender_screen')].value ? 'F' : 'M';
-
-  const isValueInRange = (value, lookupInfo) => {
-    if (!lookupInfo || lookupInfo == value) {
-      return true;
-    }
-
-    const matched = lookupInfo.match(/^(\d+)\s*[-~]\s*(\d+)$/);
-
-    if (matched) {
-      value = parseInt(value);
-
-      return !isNaN(value) && value >= parseInt(matched[1]) && value <= parseInt(matched[2]);
-    }
-    return false;
+  return {
+    tScore: subScaleScore,
+    outputText: null
   };
+}
 
-  for (let row of lookupTable) {
-    if ( 
-      isValueInRange(subScaleScore, row.rawScore) && 
-      isValueInRange(age, row.age) &&
-      isValueInRange(gender, row.sex)
-    ) {
-      return parseInt(row.tScore);
+export const getFinalSubScale = (responses, items, isAverage, lookupTable) => {
+  let total = 0, count = 0;
+  for (let i = 0; i < responses.length; i++) {
+    if (responses[i]) {
+      total += getScoreFromResponse(items[i], responses[i].value);
+      if (items[i].valueConstraints && items[i].valueConstraints.scoring) {
+        count++;
+      }
     }
   }
 
-  return subScaleScore;
+  const score = (isAverage ? total / Math.max(count, 1) : total);
+
+  if (lookupTable) {
+    for (let row of lookupTable) {
+      if (
+        isValueInRange(score, row.rawScore)
+      ) {
+        return {
+          rawScore: score,
+          outputText: row.outputText
+        };
+      }
+    }
+  }
+
+  return {
+    rawScore: score,
+    outputText: ''
+  }
 }