+++ app/widgets/StackedRadio/index.js
@@ -9,17 +9,17 @@ import { TooltipBox } from '../TooltipBox';
 
 const styles = StyleSheet.create({
   itemText: {
-    fontSize: 12,
+    fontSize: 11,
   },
   optionText: {
-    fontSize: 12
+    fontSize: 11
   },
   tooltip: {
     color: 'red'
   }
 });
 
-export const StackedRadio = ({ value, config, onChange, token }) => {
+export const StackedRadio = ({ value, config, onChange, token, onSelected }) => {
   const optionNumber = config.options.length;
   const optionWidth = `${Math.floor(75 / optionNumber)}%`;
   const tokenValues = [];
@@ -42,7 +42,7 @@ export const StackedRadio = ({ value, config, onChange, token }) => {
     for (let i = 0; i < config.itemList.length; i++) {
       tokenValues.push([]);
       for (let j = 0; j < config.options.length; j++) {
-        tokenValues[i].push(config.itemOptions[i * config.options.length + j].value);
+        tokenValues[i].push(config.itemOptions[i * config.options.length + j].value || 0);
       }
     }
   }
@@ -62,12 +62,13 @@ export const StackedRadio = ({ value, config, onChange, token }) => {
       onChange(currentValue);
     }
 
+    onSelected();
   };
 
   return (
     <View style={{ alignItems: 'stretch' }}>
       <ListItem
-        style={{ width: '95%' }}
+        style={{ width: '100%' }}
       >
         <View style={{ width: '25%' }}></View>
         {
@@ -75,10 +76,16 @@ export const StackedRadio = ({ value, config, onChange, token }) => {
             <View style={{ width: optionWidth }}>
               {option.description ? (
                 <TooltipBox text={option.description}>
-                  <Text style={styles.optionText}>{ option.name }<Text style={styles.tooltip}>*</Text></Text>
+                  <Text
+                    style={styles.optionText}
+                    allowFontScaling={false}
+                  >{ option.name.en }<Text style={styles.tooltip}>*</Text></Text>
                 </TooltipBox>
               ) : (
-                <Text style={styles.optionText}>{ option.name }</Text>
+                <Text
+                  style={styles.optionText}
+                  allowFontScaling={false}
+                >{ option.name.en }</Text>
               )}
               {option.image ? (
                 <Image
@@ -95,16 +102,22 @@ export const StackedRadio = ({ value, config, onChange, token }) => {
 
       {config.itemList.map((item, i) => (
         <ListItem
-          style={{ width: '95%' }}
+          style={{ width: '100%' }}
           key={i}
         >
           <View style={{ width: '25%' }}>
             {item.description ? (
               <TooltipBox text={item.description}>
-                <Text style={styles.itemText}>{ item.name }<Text style={styles.tooltip}>*</Text></Text>
+                <Text
+                  style={styles.itemText}
+                  allowFontScaling={false}
+                >{ item.name.en }<Text style={styles.tooltip}>*</Text></Text>
               </TooltipBox>
             ) : (
-              <Text style={styles.itemText}>{ item.name }</Text>
+              <Text
+                style={styles.itemText}
+                allowFontScaling={false}
+              >{ item.name.en }</Text>
             )}
             {item.image ? (
                 <Image
@@ -129,9 +142,9 @@ export const StackedRadio = ({ value, config, onChange, token }) => {
                   multipleChoice && (
                     <CheckBox
                       checked={
-                        currentValue[i].includes(`${option.name}${token ? ':'+tokenValues[i][j] : ''}`)
+                        currentValue[i].includes(`${option.name.en}${token ? ':'+tokenValues[i][j] : ''}`)
                       }
-                      onPress={() => handlePress(`${option.name}${token ? ':'+tokenValues[i][j] : ''}`, i)}
+                      onPress={() => handlePress(`${option.name.en}${token ? ':'+tokenValues[i][j] : ''}`, i)}
                       checkedIcon="check-square"
                       uncheckedIcon="square-o"
                       checkedColor={colors.primary}
@@ -139,8 +152,8 @@ export const StackedRadio = ({ value, config, onChange, token }) => {
                     />
                   ) || (
                     <CheckBox
-                      checked={currentValue[i] === `${option.name}${token ? ':'+tokenValues[i][j] : ''}`}
-                      onPress={() => handlePress(`${option.name}${token ? ':'+tokenValues[i][j] : ''}`, i)}
+                      checked={currentValue[i] === `${option.name.en}${token ? ':'+tokenValues[i][j] : ''}`}
+                      onPress={() => handlePress(`${option.name.en}${token ? ':'+tokenValues[i][j] : ''}`, i)}
                       checkedIcon="dot-circle-o"
                       uncheckedIcon="circle-o"
                       checkedColor={colors.primary}
@@ -181,4 +194,5 @@ StackedRadio.propTypes = {
   }).isRequired,
   token: PropTypes.bool,
   onChange: PropTypes.func.isRequired,
+  onSelected: PropTypes.func.isRequired,
 };