+++ CHANGELOG.md
@@ -4,9 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### [Unreleased]
+## [0.15.30] - 2020-03-31
+- Fix several bugs
+
+## [0.15.29] - 2020-03-29
+- Changed image resize mode
+
+## [0.15.28] - 2020-03-28
+- Fixed question name in data tab
+
+## [0.15.27] - 2020-03-21
+- Fixed EMA next button issue
+
+## [0.15.26] - 2020-03-25
+- Modified audioImageRecord widget
 
 ## [0.15.25] - 2020-03-21
-- Fixed EMA next button issue
+- Fixed next button in EMA applet
+- Removed white space above items
+- Fixed vertical offset
 
 ## [0.15.24] - 2020-03-18
 - Fixed async storage cache issue